Microphone cleanup (#1076)
* remove usless files

* split micrphone impl

* move to header

* typos

* guard headers

* refactor

* simplifying

* fixing order of includes
