--- libs/microphone/ns.ts
@@ -1,4 +1,0 @@
-
-//% color="#B4009E" weight=98 icon="\uf192"
-namespace input {
-}
